requirements.txt
streamlit==1.32.0
pywebview==4.4.1
PyGithub==2.1.1
google-generativeai==0.3.2
pyinstaller==6.1.0